Refactor simulation code into its own package
