+++ triggers/tests/BUILD.bazel
@@ -246,8 +246,8 @@ da_scala_library(
 da_scala_library(
     name = "trigger-simulation-lib",
     srcs = [
-        "src/test/scala/com/digitalasset/daml/lf/engine/trigger/CatGenerators.scala",
-        "src/test/scala/com/digitalasset/daml/lf/engine/trigger/TriggerRuleSimulationLib.scala",
+        "src/test/scala/com/digitalasset/daml/lf/engine/trigger/simulation/CatGenerators.scala",
+        "src/test/scala/com/digitalasset/daml/lf/engine/trigger/simulation/TriggerRuleSimulationLib.scala",
     ],
     scala_deps = [
         "@maven//:com_typesafe_akka_akka_stream",
@@ -277,7 +277,7 @@ da_scala_library(
 da_scala_test_suite(
     name = "trigger-simulation-lib-tests",
     timeout = "long",
-    srcs = glob(["src/test/scala/com/digitalasset/daml/lf/engine/trigger/*Test.scala"]),
+    srcs = glob(["src/test/scala/com/digitalasset/daml/lf/engine/trigger/simulation/*Test.scala"]),
     data = [
         ":acs-1.dev.dar",
         "//test-common/test-certificates",
